ref(onboarding): Split capacitor onboarding doc into multiple files
