+++ static/app/gettingStartedDocs/capacitor/capacitor/onboarding.spec.tsx
@@ -1,7 +1,8 @@
 import {renderWithOnboardingLayout} from 'sentry-test/onboarding/renderWithOnboardingLayout';
 import {screen} from 'sentry-test/reactTestingLibrary';
 
-import docs, {SiblingOption} from './capacitor';
+import {SiblingOption} from './utils';
+import docs from '.';
 
 describe('capacitor onboarding docs', () => {
   it('renders docs correctly', () => {
